@@ -4,49 +4,48 @@ import 'package:flutter/material.dart';
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Rounded Loading Button Demo',
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(title: 'Rounded Loading Button Demo'),
+      home: MyHomePage(),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateMixin {
-
-  AnimationController _animationController;
-  Animation<double> _animation;
+class _MyHomePageState extends State<MyHomePage>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _animationController;
+  late Animation<double> _animation;
 
   @override
-  void initState() { 
+  void initState() {
     super.initState();
-    
-    _animationController = AnimationController(vsync: this, duration: Duration(seconds: 1));
-    
-    _animation = new Tween<double>(begin: 0, end: 1).animate(new CurvedAnimation(parent: _animationController, curve: Curves.easeInOutCirc));
+
+    _animationController = AnimationController(
+      vsync: this,
+      duration: Duration(seconds: 1),
+    );
+
+    _animation = new Tween<double>(begin: 0, end: 1).animate(
+      new CurvedAnimation(
+        parent: _animationController,
+        curve: Curves.easeInOutCirc,
+      ),
+    );
   }
 
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
-      appBar: AppBar(
-        title: Text("Animated Check Example"),
-      ),
+      appBar: AppBar(title: Text("Animated Check Example")),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -55,17 +54,19 @@ class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateM
               child: AnimatedCheck(
                 progress: _animation,
                 size: 200,
-              )
+              ),
             ),
-            FlatButton(
-              child: Text("Check"),
-              onPressed: _animationController.forward),
-            FlatButton(
+            ElevatedButton(
+              child: Text("Forward"),
+              onPressed: _animationController.forward,
+            ),
+            ElevatedButton(
               child: Text("Reset"),
-              onPressed: _animationController.reset)
+              onPressed: _animationController.reset,
+            )
           ],
         ),
-      )
+      ),
     );
   }
 }
