@@ -1,58 +1,52 @@
 import 'dart:ui';
-
 import 'package:flutter/material.dart';
 
 class AnimatedCheck extends StatefulWidget {
-  final Animation<double> progress;
-
-  // The size of the checkmark
-  final double size;
-
-  // The primary color of the checkmark
-  final Color color;
-
-  // The width of the checkmark stroke
-  final double strokeWidth;
-
   AnimatedCheck({
-    @required this.progress,
-    @required this.size,
+    Key? key,
+    required this.progress,
+    required this.size,
     this.color,
-    this.strokeWidth});
+    this.strokeWidth,
+  }) : super(key: key);
+
+  final Animation<double> progress;
+  final double size; // The size of the checkmark
+  final Color? color; // The color of the checkmark
+  final double? strokeWidth; // The width of the checkmark's stroke
 
   @override
   State<StatefulWidget> createState() => AnimatedCheckState();
 }
 
-class AnimatedCheckState extends State<AnimatedCheck> with SingleTickerProviderStateMixin {
-
-  @override
-  void initState() { 
-    super.initState();
-  }
-
+class AnimatedCheckState extends State<AnimatedCheck>
+    with SingleTickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-
     return CustomPaint(
-      foregroundPainter: AnimatedPathPainter(widget.progress, widget.color ?? theme.primaryColor, widget.strokeWidth),
+      foregroundPainter: AnimatedPathPainter(
+        widget.progress,
+        widget.color ?? Theme.of(context).primaryColor,
+        widget.strokeWidth,
+      ),
       child: new SizedBox(
         width: widget.size,
         height: widget.size,
-      )
+      ),
     );
   }
 }
 
 class AnimatedPathPainter extends CustomPainter {
+  AnimatedPathPainter(
+    this._animation,
+    this._color,
+    this.strokeWidth,
+  ) : super(repaint: _animation);
+
   final Animation<double> _animation;
-
   final Color _color;
-
-  final double strokeWidth;
-
-  AnimatedPathPainter(this._animation, this._color, this.strokeWidth) : super(repaint: _animation);
+  final double? strokeWidth;
 
   Path _createAnyPath(Size size) {
     return Path()
@@ -62,36 +56,30 @@ class AnimatedPathPainter extends CustomPainter {
   }
 
   Path createAnimatedPath(Path originalPath, double animationPercent) {
-    final totalLength = originalPath
-        .computeMetrics()
-        .fold(0.0, (double prev, PathMetric metric) => prev + metric.length);
-
+    final totalLength = originalPath.computeMetrics().fold(
+          0.0,
+          (double prev, PathMetric metric) => prev + metric.length,
+        );
     final currentLength = totalLength * animationPercent;
-
     return extractPathUntilLength(originalPath, currentLength);
   }
 
   Path extractPathUntilLength(Path originalPath, double length) {
     var currentLength = 0.0;
-
     final path = new Path();
-
     var metricsIterator = originalPath.computeMetrics().iterator;
 
     while (metricsIterator.moveNext()) {
       var metric = metricsIterator.current;
-
       var nextLength = currentLength + metric.length;
-
       final isLastSegment = nextLength > length;
+
       if (isLastSegment) {
         final remainingLength = length - currentLength;
         final pathSegment = metric.extractPath(0.0, remainingLength);
-
         path.addPath(pathSegment, Offset.zero);
         break;
       } else {
-
         final pathSegment = metric.extractPath(0.0, metric.length);
         path.addPath(pathSegment, Offset.zero);
       }
@@ -105,10 +93,9 @@ class AnimatedPathPainter extends CustomPainter {
   @override
   void paint(Canvas canvas, Size size) {
     final animationPercent = this._animation.value;
-
     final path = createAnimatedPath(_createAnyPath(size), animationPercent);
-
     final Paint paint = Paint();
+
     paint.color = _color;
     paint.style = PaintingStyle.stroke;
     paint.strokeWidth = strokeWidth ?? size.width * 0.06;
@@ -117,7 +104,5 @@ class AnimatedPathPainter extends CustomPainter {
   }
 
   @override
-  bool shouldRepaint(CustomPainter oldDelegate) {
-    return true;
-  }
+  bool shouldRepaint(CustomPainter oldDelegate) => true;
 }
